updated the migrations to have markdown_id in pagesnapshot and cleanup_up article_detail handler
