+++ cli/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 tokio = { workspace = true, features = ["full"] }
-db = { path = "../db" }
+cores = { path = "../cores" }
 color-eyre = { workspace = true }
 clap = { version = "4.1", features = ["derive"] }
 sqlx.workspace = true
