+++ cli/src/display_user.rs
@@ -1,5 +1,5 @@
 use color_eyre::Result;
-use db::{users::get_user, PgPool};
+use cores::{users::get_user, PgPool};
 
 pub async fn display_users(pool: &PgPool) -> Result<()> {
     let users = get_user(pool).await?;
