+++ cores/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "db"
+name = "cores"
 version = "0.0.0"
 edition = "2021"
