+++ cores/src/openai_utils.rs
