+++ web/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-db = { path = "../db" }
+cores = { path = "../cores" }
 tokio = { workspace = true, features = ["full"] }
 cja = { version = "0.0.0", git = "https://github.com/coreyja/coreyja.com", branch = "main" }
 color-eyre = "0.6.3"
