+++ web/src/jobs/process_article.rs
@@ -5,12 +5,13 @@ use miette::IntoDiagnostic;
 use url::Url;
 use uuid::Uuid;
 
-use db::urls::{clean_raw_html, download_raw_html, generate_and_store_summary, persist_article, store_markdown, Markdown, Page};
+use cores::markdown::store_markdown;
+use cores::page_snapshot::{clean_raw_html, download_raw_html, PageSnapShot};
+use cores::urls::{generate_and_store_summary, Page};
 
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize)]
 pub struct ProcessArticle {
     pub page_id: Uuid,
-    pub markdown_id: Uuid,
 }
 
 #[async_trait::async_trait]
@@ -25,7 +26,30 @@ impl Job<AppState> for ProcessArticle {
             .await
             .into_diagnostic()?;
 
-        persist_article(db, page).await.unwrap();
+        let raw_html = download_raw_html(&page.url).await.unwrap();
+        let url_parsed = Url::parse(&page.url).unwrap();
+        let current_time = chrono::Utc::now();
+        let cleaned_html = clean_raw_html(&raw_html, &url_parsed).unwrap();
+
+        let page_snapshot = sqlx::query_as!(
+            PageSnapShot,
+            "INSERT INTO PageSnapShot (raw_html, fetched_at, cleaned_html, page_id) 
+        VALUES ($1, $2, $3, $4) RETURNING *",
+            raw_html,
+            current_time,
+            cleaned_html,
+            page.page_id
+        )
+        .fetch_one(db)
+        .await
+        .unwrap();
+
+        let markdown = store_markdown(db, page_snapshot).await.unwrap();
+        let markdown_id = markdown.markdown_id;
+
+        generate_and_store_summary(db, markdown_id, &cleaned_html)
+            .await
+            .unwrap();
 
         Ok(())
     }
