+++ web/src/pages.rs
@@ -2,7 +2,7 @@ use axum::{
     extract::{Path, State},
     response::{IntoResponse, Response},
 };
-use db::{urls::Markdown, users::User};
+use cores::{markdown::Markdown, page_snapshot::PageSnapShot, urls::Page, users::User};
 
 use crate::{
     templates::{Template, TemplatedPage},
@@ -52,25 +52,34 @@ pub async fn article_detail(
     Path(article_id): Path<Uuid>,
     State(state): State<AppState>,
 ) -> WebResult<Response> {
-    info!("Fetching article MD ID: {}", article_id);
+    info!("Fetching article_ID: {}", article_id);
 
-    let article = sqlx::query_as!(
-        Markdown,
-        "SELECT * FROM markdown WHERE markdown_id = $1",
-        article_id
-    )
-    .fetch_one(&state.db)
-    .await?;
+    let article = sqlx::query_as!(Page, "SELECT * FROM pages WHERE page_id = $1", article_id)
+        .fetch_one(&state.db)
+        .await?;
 
-    info!("Fetched MD: {:?}", article);
-
-    let markdown = sqlx::query_as!(
-        Markdown,
-        "SELECT * FROM Markdown WHERE markdown_id = $1",
-        article.markdown_id
+    let page_snapshot = sqlx::query_as!(
+        PageSnapShot,
+        "SELECT * FROM pagesnapshot WHERE page_id = $1 ORDER BY fetched_at DESC LIMIT 1",
+        article.page_id
     )
     .fetch_optional(&state.db)
     .await?;
+
+    let markdown = if let Some(page_snapshot) = page_snapshot {
+        let markdown = sqlx::query_as!(
+            Markdown,
+            "SELECT * FROM markdown WHERE page_snapshot_id = $1",
+            page_snapshot.page_snapshot_id
+        )
+        .fetch_optional(&state.db)
+        .await?;
+
+        markdown
+    } else {
+        None
+    };
+    info!("Fetched Article: {:?}", article);
 
     info!("Fetched markdown MD: {:?}", markdown);
