+++ web/src/users/add_url.rs
@@ -8,10 +8,10 @@ use url::Url;
 
 use crate::{jobs::process_article::ProcessArticle, AppState, WebResult};
 
-use db::{
-    urls::{
-        add_url, clean_raw_html, download_raw_html, generate_and_store_summary, store_markdown,
-    },
+use cores::{
+    markdown::store_markdown,
+    page_snapshot::{clean_raw_html, download_raw_html},
+    urls::add_url,
     users::User,
 };
 use serde::Deserialize;
@@ -27,36 +27,20 @@ pub async fn insert_article_handler(
     user: User,
     Form(form): Form<ArticleForm>,
 ) -> WebResult<impl IntoResponse> {
-    info!("Received request to insert article: {}", form.url); 
+    info!("Received request to insert article: {}", form.url); // Log the received URL
 
     let url = form.url;
     let user_id = user.user_id;
 
     let page = add_url(&state.db, &url, user_id, &true).await?;
-    let page_id = page.page().page_id;
+    let page_id: uuid::Uuid = page.page().page_id;
 
-    let raw_html = download_raw_html(&url).await?;
-    let url_parsed = Url::parse(&url)?;
-    let cleaned_html = clean_raw_html(&raw_html, &url_parsed)?;
-
-    let markdown = store_markdown(&state.db, page_id, &cleaned_html)
-        .await
-        .unwrap();
-    let markdown_id = markdown.markdown_id;
-
-    generate_and_store_summary(&state.db, markdown_id, &cleaned_html)
-        .await
-        .unwrap();
-
-    let process_article = ProcessArticle {
-        page_id,
-        markdown_id,
-    };
+    let process_article = ProcessArticle { page_id };
     process_article
         .enqueue(state.clone(), "insert_article_handler".to_string())
         .await?;
 
     Ok(Redirect::to(&format!(
-        "/articles/{markdown_id}?flash[success]=Article added"
+        "/articles/{page_id}?flash[success]=Article added"
     )))
 }
